chore: operationgenerator cleanup
