+++ packages/swagger-faker/src/OperationGenerator.test.tsx
@@ -39,8 +39,8 @@ describe('OperationGenerator', async () => {
     const operation = oas.operation('/pets', 'get')
     const operationShowById = oas.operation('/pets/{petId}', 'get')
 
-    const files = await og.get(operation, og.getSchemas(operation), options) as KubbFile.File[]
-    const getShowByIdFiles = await og.get(operationShowById, og.getSchemas(operationShowById), options) as KubbFile.File[]
+    const files = await og.get(operation, options) as KubbFile.File[]
+    const getShowByIdFiles = await og.get(operationShowById, options) as KubbFile.File[]
 
     files.forEach(file => {
       expect(FileManager.getSource(file)).toMatchSnapshot()
@@ -75,8 +75,8 @@ describe('OperationGenerator', async () => {
     const operation = oas.operation('/pets', 'get')
     const operationShowById = oas.operation('/pets/{petId}', 'get')
 
-    const files = await og.get(operation, og.getSchemas(operation), options) as KubbFile.File[]
-    const getShowByIdFiles = await og.get(operationShowById, og.getSchemas(operationShowById), options) as KubbFile.File[]
+    const files = await og.get(operation, options) as KubbFile.File[]
+    const getShowByIdFiles = await og.get(operationShowById, options) as KubbFile.File[]
 
     files.forEach(file => {
       expect(FileManager.getSource(file)).toMatchSnapshot()
@@ -109,7 +109,7 @@ describe('OperationGenerator', async () => {
       },
     )
     const operation = oas.operation('/pets', 'post')
-    const files = await og.post(operation, og.getSchemas(operation), options) as KubbFile.File[]
+    const files = await og.post(operation, options) as KubbFile.File[]
 
     files.forEach(file => {
       expect(FileManager.getSource(file)).toMatchSnapshot()
@@ -138,7 +138,7 @@ describe('OperationGenerator', async () => {
       },
     )
     const operation = oas.operation('/pet/{petId}', 'delete')
-    const files = await og.delete(operation, og.getSchemas(operation), options) as KubbFile.File[]
+    const files = await og.delete(operation, options) as KubbFile.File[]
 
     files.forEach(file => {
       expect(FileManager.getSource(file)).toMatchSnapshot()
