+++ packages/swagger-msw/src/OperationGenerator.tsx
@@ -2,11 +2,11 @@ import { createRoot } from '@kubb/react'
 import { OperationGenerator as Generator } from '@kubb/swagger'
 import { Oas } from '@kubb/swagger/components'
 
-import { Handlers } from './components/Handlers.tsx'
 import { Mock } from './components/Mock.tsx'
+import { Operations } from './components/Operations.tsx'
 
 import type { AppContextProps } from '@kubb/react'
-import type { OperationMethodResult, OperationsByMethod, OperationSchemas } from '@kubb/swagger'
+import type { OperationMethodResult, OperationsByMethod } from '@kubb/swagger'
 import type { Operation } from '@kubb/swagger/oas'
 import type { FileMeta, PluginOptions } from './types.ts'
 
@@ -17,18 +17,14 @@ export class OperationGenerator extends Generator<PluginOptions['resolvedOptions
     const root = createRoot<AppContextProps<PluginOptions['appMeta']>>({ logger: pluginManager.logger })
 
     const templates = {
-      handlers: Handlers.templates,
+      operations: Operations.templates,
       mock: Mock.templates,
       ...this.options.templates,
     }
 
-    if (!templates?.handlers) {
-      return []
-    }
-
     root.render(
       <Oas oas={oas} operations={operations} getSchemas={(...props) => this.getSchemas(...props)}>
-        <Handlers.File name="handlers" operationsByMethod={operationsByMethod} templates={templates.handlers} />
+        {templates?.operations && <Operations.File operationsByMethod={operationsByMethod} templates={templates.operations} />}
       </Oas>,
       { meta: { pluginManager, plugin } },
     )
@@ -36,13 +32,13 @@ export class OperationGenerator extends Generator<PluginOptions['resolvedOptions
     return root.files
   }
 
-  async #generate(operation: Operation, schemas: OperationSchemas, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
+  async #generate(operation: Operation, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
     const { oas, pluginManager, plugin } = this.context
 
     const root = createRoot<AppContextProps<PluginOptions['appMeta']>>({ logger: pluginManager.logger })
 
     const templates = {
-      handlers: Handlers.templates,
+      handlers: Operations.templates,
       mock: Mock.templates,
       ...options.templates,
     }
@@ -63,21 +59,21 @@ export class OperationGenerator extends Generator<PluginOptions['resolvedOptions
     return root.files
   }
 
-  async get(operation: Operation, schemas: OperationSchemas, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
-    return this.#generate(operation, schemas, options)
+  async get(operation: Operation, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
+    return this.#generate(operation, options)
   }
 
-  async post(operation: Operation, schemas: OperationSchemas, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
-    return this.#generate(operation, schemas, options)
+  async post(operation: Operation, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
+    return this.#generate(operation, options)
   }
 
-  async put(operation: Operation, schemas: OperationSchemas, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
-    return this.#generate(operation, schemas, options)
+  async put(operation: Operation, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
+    return this.#generate(operation, options)
   }
-  async patch(operation: Operation, schemas: OperationSchemas, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
-    return this.#generate(operation, schemas, options)
+  async patch(operation: Operation, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
+    return this.#generate(operation, options)
   }
-  async delete(operation: Operation, schemas: OperationSchemas, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
-    return this.#generate(operation, schemas, options)
+  async delete(operation: Operation, options: PluginOptions['resolvedOptions']): OperationMethodResult<FileMeta> {
+    return this.#generate(operation, options)
   }
 }
