+++ packages/swagger-msw/src/components/index.ts
@@ -1,2 +1,2 @@
-export { Handlers } from './Handlers.tsx'
 export { Mock } from './Mock.tsx'
+export { Operations } from './Operations.tsx'
