+++ packages/swagger-msw/src/types.ts
@@ -1,9 +1,9 @@
 import type { KubbFile, Plugin, PluginFactoryOptions, ResolveNameParams } from '@kubb/core'
 import type { Exclude, Include, Override, ResolvePathOptions } from '@kubb/swagger'
-import type { Handlers, Mock } from './components/index.ts'
+import type { Mock, Operations } from './components/index.ts'
 
 type Templates = {
-  handlers?: typeof Handlers.templates | false
+  operations?: typeof Operations.templates | false
   mock?: typeof Mock.templates | false
 }
