+++ packages/swagger-zod/src/OperationGenerator.test.tsx
@@ -41,8 +41,8 @@ describe('OperationGenerator', async () => {
     const operation = oas.operation('/pets', 'get')
     const operationShowById = oas.operation('/pets/{petId}', 'get')
 
-    const files = await og.get(operation, og.getSchemas(operation), options) as KubbFile.File[]
-    const getShowByIdFiles = await og.get(operationShowById, og.getSchemas(operationShowById), options) as KubbFile.File[]
+    const files = await og.get(operation, options) as KubbFile.File[]
+    const getShowByIdFiles = await og.get(operationShowById, options) as KubbFile.File[]
 
     files.forEach(file => {
       expect(FileManager.getSource(file)).toMatchSnapshot()
@@ -77,7 +77,7 @@ describe('OperationGenerator', async () => {
       },
     )
     const operation = oas.operation('/pets', 'post')
-    const files = await og.post(operation, og.getSchemas(operation), options) as KubbFile.File[]
+    const files = await og.post(operation, options) as KubbFile.File[]
 
     files.forEach(file => {
       expect(FileManager.getSource(file)).toMatchSnapshot()
@@ -108,7 +108,7 @@ describe('OperationGenerator', async () => {
       },
     )
     const operation = oas.operation('/pet/{petId}', 'delete')
-    const files = await og.delete(operation, og.getSchemas(operation), options) as KubbFile.File[]
+    const files = await og.delete(operation, options) as KubbFile.File[]
 
     files.forEach(file => {
       expect(FileManager.getSource(file)).toMatchSnapshot()
@@ -139,7 +139,7 @@ describe('OperationGenerator', async () => {
       },
     )
     const operation = oas.operation('/pet/{petId}', 'delete')
-    const files = await og.delete(operation, og.getSchemas(operation), options)
+    const files = await og.delete(operation, options)
 
     expect(files).toMatchSnapshot()
   })
